requirements.txt

apache-airflow>=2.7.0
pandas>=1.5.0
psycopg2-binary>=2.9.0
plotly>=5.0.0
requests>=2.28.0